some reorg in extension stuff
