+++ vscode-extension/.yarnrc
@@ -1,0 +1,1 @@
+--ignore-engines true
