+++ vscode-extension/modules/completion.js
