+++ vscode-extension/modules/diagnostic.js
