+++ vscode-extension/modules/hover.js
