+++ vscode-extension/modules/readconf.js
