+++ vscode-extension/package.json
@@ -1,8 +1,20 @@
 {
 	"name": "quickonf",
-	"displayName": "Quickonf",
-	"description": "Support for Quickonf configuration language",
 	"version": "0.0.1",
+	"displayName": "Quickonf",
+	"description": "Support for the Quickonf configuration language",
+	"publisher": "sebmacc",
+	"license": "MIT",
+	"qna": "https://github.com/tiramiseb/quickonf/discussions/categories/q-a",
+	"bugs": {
+		"url": "https://github.com/tiramiseb/quickonf/issues"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/tiramiseb/quickonf.git"
+	},
+	"homepage": "https://tiramiseb.github.io/quickonf/",
+	"icon": "icon.png",
 	"engines": {
 		"vscode": "^1.68.0"
 	},
@@ -12,7 +24,7 @@
 	"activationEvents": [
 		"onLanguage:quickonf"
 	],
-	"main": "src/main.js",
+	"main": "extension",
 	"contributes": {
 		"configuration": {
 			"title": "Quickonf",
@@ -24,29 +36,40 @@
 				}
 			}
 		},
-		"languages": [{
-			"id": "quickonf",
-			"aliases": ["Quickonf", "quickonf"],
-			"extensions": [".qconf"],
-			"configuration": "./language-configuration.json"
-		}],
-		"grammars": [{
-			"language": "quickonf",
-			"scopeName": "source.quickonf",
-			"path": "./syntaxes/quickonf.tmLanguage.json"
-		}],
+		"languages": [
+			{
+				"id": "quickonf",
+				"aliases": [
+					"Quickonf",
+					"quickonf"
+				],
+				"extensions": [
+					".qconf"
+				],
+				"configuration": "./language-configuration.json"
+			}
+		],
+		"grammars": [
+			{
+				"language": "quickonf",
+				"scopeName": "source.quickonf",
+				"path": "./syntaxes/quickonf.tmLanguage.json"
+			}
+		],
 		"semanticTokenTypes": [
-		{
-			"id": "invalid",
-			"description": "Invalid instruction"
-		}
+			{
+				"id": "invalid",
+				"description": "Invalid instruction"
+			}
 		],
 		"semanticTokenScopes": [
-		{
-			"scopes": {
-			"invalid": ["invalid.illegal"]
+			{
+				"scopes": {
+					"invalid": [
+						"invalid.illegal"
+					]
+				}
 			}
-		}
 		]
 	}
 }
